Add exchange api to maison screen

+++ app/src/main/java/com/example/designspectrum/data/product/Product.kt
@@ -5,14 +5,14 @@ import java.io.Serializable
 data class Product(
     var productName: String = "",
     var productDescription: String = "",
-    var productPrice: Int = 0,
+    var productPrice: Double = 0.0,
     var productCategory: String = "",
     var productCountry: String = "",
     var quantityInStock: Int = 0,
     var ordersQuantity: Int = 0,
     var productImageId: String = ""
 ) :Serializable{
-    constructor() : this("", "", 0, "", "", 0, 0, "")
+    constructor() : this("", "", 0.0, "", "", 0, 0, "")
 }
